tests/wpt/web-platform-tests/app-history/navigate-event/navigate-cross-origin-iframe-location.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<body>
<script>
async_test(t => {
  let iframe_url = new URL("resources/cross-origin-iframe-helper.html", location.href);
  iframe_url.hostname = get_host_info().REMOTE_HOST;
  let iframe = document.createElement("iframe")
  iframe.src = iframe_url;
  document.body.appendChild(iframe);

  window.onmessage = t.step_func_done(e => {
    assert_equals(e.data, "cancelable,canTransition,hashChange,");
  });

  appHistory.onnavigate = t.step_func_done(() => {
    assert_unreached("onnavigate should not have fired in source window");
  });
  window.onload = t.step_func(() => iframe.contentWindow.location = iframe_url + "#");
}, "location API on another window fires navigate event even if the target window is cross-origin");
</script>
<body>
